requirements.txt
# Core
streamlit==1.39.0
pandas==2.2.3
numpy==2.1.2
python-dotenv==1.0.1

# Dhan Trading
dhanhq==1.3.5

# AI/ML
openai==1.54.3
langchain==0.3.7
langchain-openai==0.2.8
langchain-community==0.3.5

# Technical Analysis
pandas-ta==0.3.14b
ta==0.11.0
ta-lib==0.4.32

# Visualization
plotly==5.24.1
matplotlib==3.9.2

# Async
nest-asyncio==1.6.0
aiohttp==3.10.10
websocket-client==1.8.0

# Database
sqlalchemy==2.0.35
alembic==1.13.3

# Utils
pytz==2024.2
requests==2.32.3
pyyaml==6.0.2

# Logging
loguru==0.7.2

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
